backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.8
python-multipart==0.0.6
jinja2==3.1.2
qrcode[pil]==7.4.2
Pillow==9.5.0
aiofiles==23.1.0
